Moved energy functions to utils
